// store/types/entity.d.ts
export declare type MessageEntity = {
  type: string;
  offset: number;
  length: number;
  url?: string;
  info?: string;
};
export declare type PicBaseInfo = {
  uuid: string;
  type: string;
  size: number;
  width: number;
  height: number;
  url: string;
};
export declare type AtUsersInfoItem = {
  atUserID: string;
  groupNickname: string;
};
// ...
export declare type PublicUserItem = {
  gender: number;
  nickname: string;
  userID: string;
  faceURL: string;
  ex: string;
};
// ...
export declare type MessageItem = {
  clientMsgID: string;
  serverMsgID: string;
  createTime: number;
  sendTime: number;
  sessionType: SessionType;
  sendID: string;
  recvID: string;
  msgFrom: number;
  contentType: MessageType;
  senderPlatformID: Platform;
  senderNickname: string;
  senderFaceUrl: string;
  groupID: string;
  content: string;
  seq: number;
  isRead: boolean;
  status: MessageStatus;
  isReact: boolean;
  isExternalExtensions: boolean;
  offlinePush: OfflinePush;
  attachedInfo: string;
  ex: string;
  localEx: string;
  textElem: TextElem;
  cardElem: CardElem;
  pictureElem: PictureElem;
  soundElem: SoundElem;
  videoElem: VideoElem;
  fileElem: FileElem;
  mergeElem: MergeElem;
  atTextElem: AtTextElem;
  faceElem: FaceElem;
  locationElem: LocationElem;
  customElem: CustomElem;
  quoteElem: QuoteElem;
  notificationElem: NotificationElem;
  advancedTextElem: AdvancedTextElem;
  typingElem: TypingElem;
  attachedInfoElem: AttachedInfoElem;
};
export declare type TextElem = {
  content: string;
};
export declare type CardElem = {
  userID: string;
  nickname: string;
  faceURL: string;
  ex: string;
};
export declare type AtTextElem = {
  text: string;
  atUserList: string[];
  atUsersInfo?: AtUsersInfoItem[];
  quoteMessage?: MessageItem;
  isAtSelf?: boolean;
};
export declare type NotificationElem = {
  detail: string;
};
export declare type AdvancedTextElem = {
  text: string;
  messageEntityList: MessageEntity[];
};
export declare type TypingElem = {
  msgTips: string;
};
export declare type CustomElem = {
  data: string;
  description: string;
  extension: string;
};
export declare type FileElem = {
  filePath: string;
  uuid: string;
  sourceUrl: string;
  fileName: string;
  fileSize: number;
};
export declare type FaceElem = {
  index: number;
  data: string;
};
export declare type LocationElem = {
  description: string;
  longitude: number;
  latitude: number;
};
export declare type MergeElem = {
  title: string;
  abstractList: string[];
  multiMessage: MessageItem[];
  messageEntityList: MessageEntity[];
};
export declare type OfflinePush = {
  title: string;
  desc: string;
  ex: string;
  iOSPushSound: string;
  iOSBadgeCount: boolean;
};
export declare type PictureElem = {
  sourcePath: string;
  sourcePicture: Picture;
  bigPicture: Picture;
  snapshotPicture: Picture;
};
export declare type AttachedInfoElem = {
  groupHasReadInfo: GroupHasReadInfo;
  isPrivateChat: boolean;
  isEncryption: boolean;
  inEncryptStatus: boolean;
  burnDuration: number;
  hasReadTime: number;
  notSenderNotificationPush: boolean;
  messageEntityList: MessageEntity[];
  uploadProgress: UploadProgress;
};
export declare type UploadProgress = {
  total: number;
  save: number;
  current: number;
};
export declare type GroupHasReadInfo = {
  hasReadCount: number;
  hasReadUserIDList: string[];
  groupMemberCount: number;
};
export declare type Picture = {
  uuid: string;
  type: string;
  size: number;
  width: number;
  height: number;
  url: string;
};
export declare type QuoteElem = {
  text: string;
  quoteMessage: MessageItem;
};
export declare type SoundElem = {
  uuid: string;
  soundPath: string;
  sourceUrl: string;
  dataSize: number;
  duration: number;
};
export declare type VideoElem = {
  videoPath: string;
  videoUUID: string;
  videoUrl: string;
  videoType: string;
  videoSize: number;
  duration: number;
  snapshotPath: string;
  snapshotUUID: string;
  snapshotSize: number;
  snapshotUrl: string;
  snapshotWidth: number;
  snapshotHeight: number;
};